+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 * Don't require /proc/self/pagemap to exist for savestating
 * Switch to autotools
+* More optimized Ram Search
 
 ### Fixed